a lot of bled fixes

+++ srcs/variations/ft_vsnprintf.c
@@ -6,7 +6,7 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/01 18:07:58 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/12/07 21:54:34 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/12/08 14:21:39 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ inline int	ft_vsnprintf(char *dst,
 
 	if ((out = internal_vprintf(format, ap)))
 	{
-		if (len <= out)
+		if (len <= (size_t)out)
 			out = len;
 		ft_strncpy(dst, g_buf, out);
 		dst[out] = '\0';
